add tag lookup by most specific to most general to handle record tag #r[.<recordname>]
it first looks for a handler with the full path, if not found drops the
last path item and retries until no path item is available.

it stops when a handler is found or when no more path items are
available.

in this way you can override a more generic handler with a more specific
one that only matches a specific path (for example only the records that
have the name "state")